requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
pytz>=2023.3
jpholiday>=0.1.9
pygame>=2.5.2
aiohttp>=3.8.5